[+] interface + abstract class for reminder type models;
